requirements.txt
django==3.0.6
djangorestframework==3.12.1